code for feeding selection text to the kernel
Based on `nvim.current.buffer.mark('<')`, `'>'` respectively. This
enables one to execute for example out of place commands (even visual
blocks...):

```
bla blah bla print('test') more blah
some bla     test = 5; test
```

For whatever reason we have the above example. Then it is possible to
select the text made out of `print('test')` and `test = 5; test` with a
visual block and feed it to the kernel for execution. This works because
it has the same indent level! But in the more usual case, one would
select either `print('test')` or `test = 5; test` and it will execute as
expected