some more refactoring
Experimenting with code organization until satisfactory.

Also, now use `:delmarks <>` only after a range is given, not single
line. So only after calling `JExecute` with a range object these marks
will be deleted. Not ideal, but until I find a better solution I'm
sticking with this one.

Also it's enough to verify if the line numbers sent to `JExecute` are in
the valid range (in `neovim` line numbering starts from 1) without
worrying about the column selections.